requirements.txt
ansible==7.0.0
black==22.3.0
# faasmctl>=0.34.0
myfaasmctl==0.0.20
flake8==3.9.2
hoststats==0.1.1
invoke>=2.1.0
matplotlib>=3.7.2
numpy>=1.25.0
pandas>=2.0.0
scipy>=1.12.0
